Fix PHP Lint Issue
Big into to Int.

+++ migration/migrator/data/course_tables.sql
@@ -1534,7 +1534,7 @@ CREATE TABLE public.polls (
     question_type character varying(35) DEFAULT 'single-response-multiple-correct'::character varying,
     release_histogram character varying(10) DEFAULT 'never'::character varying,
     release_answer character varying(10) DEFAULT 'never'::character varying,
-    duration bigint DEFAULT 0,
+    duration integer DEFAULT 0,
     end_date timestamp with time zone DEFAULT '1900-02-01 00:00:00-05'::timestamp with time zone NOT NULL
 );
